Formal: experiment in stubbing out dependencies
This is currently a failed and incomplete experiment.

Failed because it is far too much work to do it this way by hand.
(Should we automate or should we do something else?)

Incomplete because, at the moment, the stubs deterministically do just
one thing (which is probably incorrect, anyway).
I hit a limitation of the formal tools ability to pass arguments
to the SUT that I need to fix before continuing.